# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskFlow API",
        "description": "API REST de gestion de proyectos y tareas. Backend puro: sin vistas y sin\nsesiones, pensada para que la consuma cualquier cliente.\n\n**Autenticacion.** Registrate o inicia sesion para recibir un token, y\nmandalo en cada peticion como `Authorization: Bearer <token>`. Los tokens\ncaducan a los siete dias.\n\n**Aislamiento.** Cada usuario solo ve y edita sus propios proyectos. Pedir\nun recurso de otra persona devuelve 403; las tareas heredan el permiso del\nproyecto al que pertenecen.\n\n**Datos de prueba.** Tras `php artisan migrate:fresh --seed` existe el\nusuario `demo@taskflow.test` con la contrasena `contrasena-demo`, y un\ncuarto proyecto que pertenece a otra persona para poder comprobar el 403.",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Registra un usuario y devuelve su primer token",
                "description": "Limitado a 5 peticiones por minuto por combinacion de email e IP.",
                "operationId": "auth.register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Usuario Demo",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "demo@taskflow.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "contrasena-demo",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "contrasena-demo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos invalidos o email ya registrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Emite un token para un usuario existente",
                "description": "Devuelve el mismo 422 tanto si el email no existe como si la contrasena\nes incorrecta. Distinguirlos convertiria este endpoint en un\nverificador de que direcciones estan registradas.\n\nLimitado a 5 intentos por minuto por combinacion de email e IP.",
                "operationId": "auth.login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "demo@taskflow.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "contrasena-demo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token emitido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales incorrectas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Revoca el token usado en la peticion",
                "description": "Solo ese token. Cerrar sesion en el movil no desconecta la del navegador.",
                "operationId": "auth.logout",
                "responses": {
                    "204": {
                        "description": "Token revocado"
                    },
                    "401": {
                        "description": "Sin token o token invalido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Autenticacion"
                ],
                "summary": "Devuelve el usuario autenticado",
                "description": "Devuelve el usuario autenticado.",
                "operationId": "auth.me",
                "responses": {
                    "200": {
                        "description": "Usuario actual",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sin token o token invalido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Usuario",
                "properties": {
                    "id": {
                        "description": "El esquema OpenAPI vive junto al codigo que produce la respuesta: si un\ncampo cambia aqui y no alli, la diferencia salta a la vista en el diff.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Usuario Demo"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "demo@taskflow.test"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Error",
                "description": "Cuerpo de todo error salvo los de validacion.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Error de validacion",
                "description": "Unico cuerpo de error con estructura anidada.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "title": [
                                "The title field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "TokenResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "user": {
                                "$ref": "#/components/schemas/User"
                            },
                            "token": {
                                "type": "string",
                                "example": "3|kR7pQ2mZ..."
                            },
                            "token_type": {
                                "type": "string",
                                "example": "Bearer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token devuelto por /auth/register o /auth/login.",
                "scheme": "bearer"
            }
        }
    },
...
}